+++ keeparelease/keeparelease_test.go
@@ -6,7 +6,7 @@ import (
 	"github.com/lithammer/dedent"
 )
 
-const changelog string = `
+const semverChangelog string = `
 # Changelog
 All notable changes to this project will be documented in this file.
 
@@ -42,27 +42,82 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#2]: https://github.com/rgreinho/keeparelease/pull/2
 `
 
+const calverChangelog string = `
+# Changelog
+
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Calendar Versioning](https://calver.org/).
+
+## [Unreleased]
+
+## Added
+
+- Add a mechanism to configure the common tools using environment variables.
+
+## [19.12.1]
+
+### Added
+
+- Add the following features to docgen-nodes-reboot:
+  - Ability to skip the n first nodes.
+  - Ability to target a specific node by ID.
+  - Ability to exit on failure instead of moving forward to the next step.
+
+### Changed
+
+- Replace all the output statements by logging statements associated with the appropriate loggin level.
+
+### Fixed
+
+- Fix the AWS EC2 rate limit issue with docgen-nodes-reboot.
+`
+
 func TestParseChangelog00(t *testing.T) {
-	title, content, err := ParseChangelog(changelog)
-	if err != nil {
-		t.Fatalf("failed to parse the changelog: %s", err)
+	testcases := []struct {
+		changelog       string
+		expectedTitle   string
+		expectedContent string
+	}{
+		{semverChangelog, "0.4.0", trimEdges(dedent.Dedent(`
+    ### Added
+
+    - A feature
+
+    ### Fixed
+
+    - A fix
+    `), " \n")},
+		{calverChangelog, "19.12.1", trimEdges(dedent.Dedent(`
+    ### Added
+
+    - Add the following features to docgen-nodes-reboot:
+      - Ability to skip the n first nodes.
+      - Ability to target a specific node by ID.
+      - Ability to exit on failure instead of moving forward to the next step.
+
+    ### Changed
+
+    - Replace all the output statements by logging statements associated with the appropriate loggin level.
+
+    ### Fixed
+
+    - Fix the AWS EC2 rate limit issue with docgen-nodes-reboot.
+    `), " \n")},
 	}
-	expectedTitle := "0.4.0"
-	expectedContent := trimEdges(dedent.Dedent(`
-  ### Added
 
-  - A feature
+	for _, tc := range testcases {
+		title, content, err := ParseChangelog(tc.changelog)
+		if err != nil {
+			t.Fatalf("failed to parse the changelog: %s", err)
+		}
 
-  ### Fixed
-
-  - A fix
-  `), " \n")
-
-	if dedent.Dedent(title) != dedent.Dedent(expectedTitle) {
-		t.Fatalf("Error: title is %s, but expected is %s", title, expectedTitle)
+		if dedent.Dedent(title) != dedent.Dedent(tc.expectedTitle) {
+			t.Fatalf("title is %q, but expected is %q", title, tc.expectedTitle)
+		}
+		if content != tc.expectedContent {
+			t.Fatalf("content is %q, \nbut expected is %q.", content, tc.expectedContent)
+		}
 	}
-	if content != expectedContent {
-		t.Fatalf("Error: content is %s, \nbut expected is %s.", content, expectedContent)
-	}
-
 }