+++ CHANGELOG.md
@@ -7,11 +7,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [[1.1.1]] - 2019-12-16
+
+### Change
+
+- Replace the `Makefile` by an `Invoke` task file.
+
+### Fixed
+
+- Fix the semver regex to make if work for calver as well.
+
+## [[1.1.0]] - 2019-07-19
+
 ## [[1.0.0]] - 2019-06-23
 
 Initial release.
 
 [//]: # (Release links)
+[1.0.0]: https://github.com/rgreinho/keeparelease/releases/tag/1.1.0
+[1.1.0]: https://github.com/rgreinho/keeparelease/releases/tag/1.1.0
+[1.1.1]: https://github.com/rgreinho/keeparelease/releases/tag/1.1.1
 
 
 [//]: # (Issue/PR links)